Crate bip_handshake

Modules

transports
Built in objects implementing Transport.

Structs

CompleteMessage
Message containing completed handshaking information.

Extensions
Extensions supported by either end of a handshake.

Handshaker
Handshaker which is both Stream and Sink.

HandshakerBuilder
Build configuration for Handshaker object creation.

HandshakerConfig
Configures the internals of a Handshaker.

HandshakerSink
Sink portion of the Handshaker for initiating handshakes.

HandshakerStream
Stream portion of the Handshaker for completed handshakes.

InitiateMessage
Message used to initiate a handshake with the Handshaker.

Enums

FilterDecision
Filtering decision made for a given handshake.

Protocol
Protocol information transmitted as part of the handshake.

Traits

DiscoveryInfo
Trait for advertisement information that other peers can discover.

HandshakeFilter
Trait for filtering connections during handshaking.

HandshakeFilters
Trait for adding and removing HandshakeFilters.

LocalAddr
Trait for getting the local address.

Transport
Trait for initializing connections over an abstract Transport.

Type Definitions

InfoHash
Bittorrent InfoHash.

PeerId
Bittorrent PeerId.