Struct bip_handshake::BTHandshaker

pub struct BTHandshaker<T> where
    T: Send,  { /* fields omitted */ }

Bittorrent TCP peer handshaker.

Methods

impl<T> BTHandshaker<T> where
T: From<BTPeer> + Send + 'static,

fn new<C>(chan: C, listen: SocketAddr, pid: PeerId) -> Result<BTHandshaker<T>> where
C: Channel<T> + 'static,

Create a new BTHandshaker with the given PeerId and bind address which will forward metadata and handshaken connections onto the provided channel.

fn with_protocol<C>(
chan: C,
listen: SocketAddr,
pid: PeerId,
protocol: &'static str
) -> Result<BTHandshaker<T>> where
C: Channel<T> + 'static,

Similar to BTHandshaker::new() but allows a client to specify a custom protocol that the handshaker will specify during the handshake.

Panics if the length of the provided protocol exceeds 255 bytes.

fn register_hash(&self, hash: InfoHash)

Register interest for the given InfoHash allowing connections for the given InfoHash to succeed. Connections already in the handshaking process may not be effected by this call.

By default, a BTHandshaker will be interested in zero InfoHashs.

This is a blocking operation.

fn deregister_hash(&self, hash: InfoHash)

Deregister interest for the given InfoHash causing connections for the given InfoHash to fail. Connections already in the handshaking process may not be effected by this call.

By default, a BTHandshaker will be interested in zero InfoHashs.

This is a blocking operation.