Trait bip_handshake::Transport

pub trait Transport {
    type Socket: AsyncRead + AsyncWrite + 'static;
    type FutureSocket: Future<Item = Self::Socket, Error = Error> + 'static;
    type Listener: Stream<Item = (Self::Socket, SocketAddr), Error = Error> + LocalAddr + 'static;
    fn connect(addr: &SocketAddr, handle: &Handle) -> Result<Self::FutureSocket>;
    fn listen(addr: &SocketAddr, handle: &Handle) -> Result<Self::Listener>;
}

Trait for initializing connections over an abstract Transport.

Associated Types

type Socket: AsyncRead + AsyncWrite + 'static

Concrete socket.

type FutureSocket: Future<Item = Self::Socket, Error = Error> + 'static

Future Self::Socket.

type Listener: Stream<Item = (Self::Socket, SocketAddr), Error = Error> + LocalAddr + 'static

Concrete listener.

Required Methods

fn connect(addr: &SocketAddr, handle: &Handle) -> Result<Self::FutureSocket>

Connect to the given address over this transport, using the supplied Handle.

fn listen(addr: &SocketAddr, handle: &Handle) -> Result<Self::Listener>

Listen to the given address for this transport, using the supplied Handle.